Move StatusBar components to features directory
Refactored StatusBar and related files from src/components/StatusBar to src/features/status-bar/components for improved project structure. Updated imports in HomePage.jsx and Project.jsx to reflect the new paths.
