+++ src/features/status-bar/components/StatusBar.jsx
